feat(core-th): split theme + make more consistent
